refactor(TaskListWidget): Extract menu settings to settings file
