+++ src/widgets/task-list/settings.ts
@@ -1,3 +1,6 @@
+import { ArrowDown01, ArrowDownAZ, ArrowUp01, ArrowUpAZ } from "lucide-react"
+
+import { MenuItemGroup } from "~/components/MenuButton"
 import { SortOrder } from "~/lib/sort"
 
 import { Task } from "./data"
@@ -31,3 +34,91 @@ export const taskListSettings = createSettingsAtom<TaskListSettings>(
 )
 
 export const useTaskListSettings = taskListSettings.useSettings
+
+const sortIcons = {
+  label: { asc: ArrowDownAZ, desc: ArrowUpAZ },
+  checked: { asc: ArrowDown01, desc: ArrowUp01 },
+  createdAt: { asc: ArrowDown01, desc: ArrowUp01 },
+}
+
+export const getMenuSettings = (
+  id: string,
+  settings: TaskListSettings
+): MenuItemGroup[] => {
+  const { sort } = settings
+
+  const changeSort = (key: keyof Task) => {
+    const order = sort?.key === key && sort.order === "asc" ? "desc" : "asc"
+    taskListSettings.setOption(id, "sort", { key: key, order })
+  }
+
+  const getIconByKey = (key: keyof typeof sortIcons) =>
+    sort?.key !== key ? undefined : sortIcons[key][sort.order]
+
+  return [
+    {
+      label: "Behavior",
+      items: [
+        {
+          label: "Compact list",
+          keepOpen: true,
+          selectable: {
+            checked: settings.compactList,
+            onChange: checked =>
+              taskListSettings.setOption(id, "compactList", checked),
+          },
+        },
+        {
+          label: "Hide selected",
+          keepOpen: true,
+          selectable: {
+            checked: settings.hideChecked,
+            onChange: checked =>
+              taskListSettings.setOption(id, "hideChecked", checked),
+          },
+        },
+        {
+          label: "No line wrap",
+          keepOpen: true,
+          selectable: {
+            checked: settings.noWrap,
+            onChange: checked =>
+              taskListSettings.setOption(id, "noWrap", checked),
+          },
+        },
+        {
+          label: "No line wrap (checked)",
+          keepOpen: true,
+          selectable: {
+            checked: settings.checkedNoWrap,
+            onChange: checked =>
+              taskListSettings.setOption(id, "checkedNoWrap", checked),
+          },
+        },
+      ],
+    },
+    {
+      label: "Sort",
+      items: [
+        {
+          label: "Checked",
+          icon: getIconByKey("checked"),
+          keepOpen: true,
+          onClick: () => changeSort("checked"),
+        },
+        {
+          label: "Alphabetically",
+          icon: getIconByKey("label"),
+          keepOpen: true,
+          onClick: () => changeSort("label"),
+        },
+        {
+          label: "Created at",
+          icon: getIconByKey("createdAt"),
+          keepOpen: true,
+          onClick: () => changeSort("createdAt"),
+        },
+      ],
+    },
+  ]
+}
